Rename and move leading components to topic
