+++ src/components/topic/banner/leading-video.js
@@ -2,8 +2,8 @@ import { Waypoint } from 'react-waypoint'
 import { replaceGCSUrlOrigin } from '@twreporter/core/lib/utils/storage-url-processor'
 import PropTypes from 'prop-types'
 import React from 'react'
-import SoundMuteIcon from '../../../static/asset/sound-mute.svg'
-import SoundOnIcon from '../../../static/asset/sound-on.svg'
+import SoundMuteIcon from '../../../../static/asset/sound-mute.svg'
+import SoundOnIcon from '../../../../static/asset/sound-on.svg'
 import styled, { css } from 'styled-components'
 
 const Container = styled.div`
@@ -12,19 +12,15 @@ const Container = styled.div`
   height: ${props => props.viewportHeight};
 `
 
-const topicVideoStyle = `
+const Video = styled.video`
+  display: block;
+  width: 100%;
+  height: 100%;
   position: absolute;
   top: 50%;
   left: 50%;
   transform: translate(-50%,-50%);
   object-fit: cover;
-  height: 100%;
-`
-
-const Video = styled.video`
-  display: block;
-  width: 100%;
-  ${props => props.isTopic ? topicVideoStyle : ''}
 `
 
 const VideoMask = styled.div`
@@ -107,7 +103,7 @@ class LeadingVideo extends React.PureComponent {
   }
 
   render() {
-    const { filetype, loop, poster, src, title, viewportHeight, isTopic } = this.props
+    const { filetype, loop, poster, src, title, viewportHeight } = this.props
     const { isMuted } = this.state
 
     // On the mobile devices (iOS 10 above),
@@ -134,7 +130,6 @@ class LeadingVideo extends React.PureComponent {
             autoPlay
             muted={isMuted}
             loop={loop}
-            isTopic={Boolean(isTopic)}
           >
             <source src={replaceGCSUrlOrigin(src)} type={filetype} />
           </Video>
@@ -160,8 +155,7 @@ LeadingVideo.propTypes = {
   poster: PropTypes.string,
   src: PropTypes.string,
   title: PropTypes.string,
-  viewportHeight: PropTypes.string,
-  isTopic: PropTypes.bool
+  viewportHeight: PropTypes.string
 }
 
 LeadingVideo.defaultProps = {
@@ -171,8 +165,7 @@ LeadingVideo.defaultProps = {
   poster: '',
   src: '',
   title: '',
-  viewportHeight: '100vh',
-  isTopic: false
+  viewportHeight: '100vh'
 }
 
 export default LeadingVideo
